Semi-reverted the heat changes. Temperature once again uses floats instead of shorts. The shorts ended up being harder to understand. However many of the aspects which unified item temp handling have remained, so overall it's a net positive.

+++ TFC_Shared/src/TFC/Food/ItemFoodTFC.java
@@ -95,7 +95,7 @@ public class ItemFoodTFC extends ItemTerra implements ISize, IFood
 
 			if(TFC_ItemHeat.HasTemp(is))
 			{
-				int meltTemp = TFC_ItemHeat.IsCookable(is);
+				float meltTemp = TFC_ItemHeat.IsCookable(is);
 				if(meltTemp != -1)
 					arraylist.add(TFC_ItemHeat.getHeatColorFood(TFC_ItemHeat.GetTemp(is), meltTemp));
 			}
